Extract CityGML terrain conformer
